refactor: mark server-only entries via server-only import instead of a separate plugin
Per review feedback, drop the standalone server-only-guard plugin and
instead build on the boundary-modules plugin from #2167: the public
@solidjs/start/http and @solidjs/start/middleware entries now carry an
import "server-only" marker, so pulling them into the client bundle
fails at resolve time with an actionable error (#2068).

- extract boundary-modules into its own module and extend its errors
  with fix guidance ("use server" / clientOnly())
- split src/http into a marked public entry re-exporting ./http.ts, so
  the isomorphic <HttpHeader> (which guards usage behind isServer) can
  keep importing the helpers without tripping the client-env check
- @solidjs/start/config is no longer guarded: it is loaded directly by
  Node when evaluating vite.config.ts, outside the plugin pipeline, so
  it cannot carry the marker
