+++ .changeset/fix-2068-server-only-guard.md
@@ -2,4 +2,4 @@
 "@solidjs/start": patch
 ---
 
-fix: fail loudly when client-reachable code imports `@solidjs/start`'s server-only entry points (`/http`, `/middleware`, `/config`) instead of silently shipping them to the browser, where they crashed hydration and broke unrelated actions/forms with no diagnostic (#2068)
+fix: mark `@solidjs/start/http` and `@solidjs/start/middleware` as `server-only` so importing them from client-reachable code fails loudly at dev/build time, instead of silently shipping them to the browser where they crashed hydration and broke unrelated actions/forms with no diagnostic (#2068)
